// src/domain/enums/HazardType.ts
// commented types are returned by the API but are not shown in the current version
export enum HazardType {
  // ACCIDENT = 'ACCIDENT',
  AVALANCHE = 'AVALANCHE',
  BIOMEDICAL = 'BIOMEDICAL',
  // CIVILUNREST = 'CIVILUNREST',
  // COMBAT = 'COMBAT',
  CYCLONE = 'CYCLONE',
  DROUGHT = 'DROUGHT',
  EARTHQUAKE = 'EARTHQUAKE',
  // EQUIPMENT = 'EQUIPMENT',
  EXTREMETEMPERATURE = 'EXTREMETEMPERATURE',
  FLOOD = 'FLOOD',
  HIGHSURF = 'HIGHSURF',
  HIGHWIND = 'HIGHWIND',
  // INCIDENT = 'INCIDENT',
  LANDSLIDE = 'LANDSLIDE',
  MANMADE = 'MANMADE',
  MARINE = 'MARINE',
  // SEVEREWEATHER = 'SEVEREWEATHER',
  STORM = 'STORM',
  // TERRORISM = 'TERRORISM',
  VOLCANO = 'VOLCANO',
  WILDFIRE = 'WILDFIRE',
  WINTERSTORM = 'WINTERSTORM',
  TORNADO = 'TORNADO',
  TSUNAMI = 'TSUNAMI',
}